refactor dummy config
